python/requirements.txt
python-socketio
eventlet
greenlet
